packages/backend/requirements.txt
fastapi==0.112.2
psycopg2-binary==2.9.9
boto3==1.35.11
uvicorn==0.30.6
python-dotenv==1.0.1
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
h11==0.14.0
idna==3.8
pydantic==2.8.2
pydantic_core==2.20.1
sniffio==1.3.1
starlette==0.38.4
typing_extensions==4.12.2
SQLAlchemy==2.0.34
asyncpg==0.29.0
websockets==13.0.1